Rename logging to core_logging.
