+++ examples/direct/bubble_example.py
@@ -15,7 +15,7 @@ from smarts.core.local_traffic_provider import LocalTrafficProvider
 from smarts.core.observations import Observation
 from smarts.core.scenario import Scenario
 from smarts.core.smarts import SMARTS
-from smarts.core.utils.logging import timeit
+from smarts.core.utils.core_logging import timeit
 from smarts.sstudio.sstypes import Bubble, PositionalZone, SocialAgentActor
 from smarts.zoo.agent_spec import AgentSpec
 from smarts.zoo.registry import register
