+++ smarts/ray/sensors/tests/test_ray_sensor_resolver.py
@@ -30,7 +30,7 @@ from smarts.core.simulation_frame import SimulationFrame
 from smarts.core.simulation_local_constants import SimulationLocalConstants
 from smarts.core.smarts import SMARTS
 from smarts.core.sumo_traffic_simulation import SumoTrafficSimulation
-from smarts.core.utils.logging import diff_unpackable
+from smarts.core.utils.core_logging import diff_unpackable
 from smarts.ray.sensors.ray_sensor_resolver import RaySensorResolver
 
 AGENT_IDS = [f"agent-00{i}" for i in range(3)]
